new attr: 'picard_param_name' for picard tools
A common case with params for picard commands is to pass them through
verbatim to picard with perhaps a slight transformation of the param
name (e.g., input_file=foo.bam => INPUT=foo.bam). The intent of this
change is to allow specifying params like so:

has_input => [
...
    input_file => {
        ...
        picard_param_name => 'INPUT',
    }
],

This enables us to use the reflection capabilities of UR to
automatically generate the picard command line args in many cases (of
course there's a hook for times when extra processing is necessary).

The next commit will include the implementation of the behaviour
described in the previous paragraph :)